requirements.txt
requests
pandas
fastparquet
numpy
matplotlib
seaborn
dotenv
tqdm
ipykernel
geopandas
shapely
scikit-learn
pandas
fastapi
uvicorn
scikit-learn
pyyaml
alphashape
pyarrow